Build set() method
This also covers cases of appending a new node to the tail of a linked
list if the obtained array index already contains a node. However,
the method does not check whether buckets need to be doubled because,
IMHO, this requires the length() method to calculate the current load
factor and the values() method to actually build a new array of double
size.